R2B: make recipe settings lint-clean

+++ crates/network/crucible-target-26-2/src/r2b_recipe.rs
@@ -41,7 +41,7 @@ impl RecipeBookSettingFlags {
     }
 
     const fn wire_bool(self, bit: u32) -> u8 {
-        ((self.0 >> bit) & 1) as u8
+        (self.0 >> bit) & 1
     }
 }
